old-tests/submission/W3C/domhtml/HTMLScriptElement05.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>script</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-script-element.html#dom-script-type" rel="help">
<link rel="author" title="W3C">
</head>
<body>

<script type="application/javascript">

  var __domtest = 0;

</script>

<script>

var domtest = async_test("The IDL attributes src, type, charset, and defer, each must reflect the respective content attributes of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vtype;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("script");
      assert_equals(nodeList.length,4);
      testNode = nodeList.item(2);
      vtype = testNode.type;

      assert_equals(vtype,"application/javascript");
       
    domtest.done();
}));

</script>
<p>The IDL attributes src, type, charset, and defer, each must reflect the respective content attributes of the same name.</p>
<div id="log"></div>
</body></html>
